requirements.txt
flask
tldextract
python-whois
dnspython
requests
gunicorn
